First commit. This project is an express.js project as backend, and an angular project as a client is the browser. Angular version 5+, is written in typescript, used the angular cli too. Third party angular component library is the primeNG component suit.
The app gooal:
Be an change log editor, as the name shows.
We can use it for more project change log editor.
First page is used to select the program, then we can add new change.
In the program we can add new program, to write the changes. We define the program changes directory path, name.

The angular build make the client, then the node .\server.js start the web server. It runs on the 3333 port default.